requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# Server/app/__init__.py: 1
# Server/app/blueprints.py: 1
# Server/app/hooks/error.py: 1
# Server/app/views/__init__.py: 4
# Server/app/views/sample/sample.py: 2
Flask == 1.0.2

# Server/app/extensions.py: 1
Flask_Cors == 3.0.6

# Server/app/extensions.py: 2
Flask_JWT_Extended == 3.13.1

# Server/app/views/__init__.py: 5
# Server/app/views/sample/__init__.py: 1
Flask_RESTful == 0.3.6

# Server/app/__init__.py: 2
Werkzeug == 0.14.1

# Server/app/extensions.py: 4
# Server/app/views/sample/sample.py: 1
flasgger == 0.9.1

# Server/app/extensions.py: 3
flask_mongoengine == 0.9.5
